Add pkgconfig, numpy and cython to setup-requires
