+++ setup.cfg
@@ -1,0 +1,4 @@
+[metadata]
+setup-requires = pkgconfig
+    numpy
+    cython
